refactor(app): one TV pairing form for box and box-less, driven by a seam
The box-side SmartTvSetup and the remote-only DirectTvSetup had grown into two
different pairing screens with two different Google TV flows — the "grow two of
something" mistake this collapses. The tester was right: the box screen's
compact brand picker is nicer, and there is no reason the box-less path should
not use it.

WHY IT WASN'T literal reuse: every SmartTvSetup pairing action is an api.tv*
POST to the box agent, and in remote-only mode there is no box to POST to. So
the fix is a driver seam, not a copy:

- components/TvPairForm — the shared form (brand picker, LAN scan, IP + optional
  MAC, the Android TV code step, the result line). Lifted verbatim from
  SmartTvSetup, so its hard-won logic — identify-then-act, the blocked/force-
  pair escape hatch, the settled-vs-narration message split — is unchanged.
- lib/tvPairDriver — the TvPairDriver interface + boxDriver, which forwards every
  action to the same api.* SmartTvSetup already called. Box behaviour is
  byte-identical.
- lib/tvdirect/directDriver — the box-less driver: pairs straight from the phone
  via the tvdirect modules and persists to the on-phone store. It exposes only
  the brands the app can reach without a box (Roku, Google TV), so the SAME
  brand picker shows five with a box and two without — no brand-specific UI
  branching. It also LAN-gates the host before contacting it.

SmartTvSetup keeps only its box chrome (the "Connected: <adapter>" row and the
multi-TV backend picker, both fed by /api/tv). DirectTvSetup keeps only the
mode toggle and the TV list. components/AndroidTvPairCard is DELETED — its flow
is now the shared form's Android TV branch.

Verified against REAL hardware through the one form:
- BOX path, against the live box (lenovodesktop, agent 2.9.70, its real LG
  webOS TV): the connected row shows "LG webOS (10.7.0.205)", the backend picker
  shows webos/lgcom, all five brand chips render, and "Scan for TVs" round-trips
  boxDriver -> api.tvDiscover -> the box (returned {tvs:[]}, the LG is already
  paired) and the form shows "No TVs found". No console errors.
- DIRECT path, against a stub Roku: typing the IP and pressing "Add Roku" ran
  directDriver.pair -> rokuIdentify (hit the stub) -> store, the form showed
  "Connected: Harness Roku", and the TV persisted. remoteOnly correctly stayed
  OFF because a box is present (auto-enable is empty-fleet only).
- tsc clean; app-input suite 198/198.

NOT verified: directDriver has no bare-Node unit test (it imports the native
socket via atvnative, unloadable in the Node runner) — the pieces it composes
are tested and the glue is harness-proven. The Android TV direct CODE step
specifically was exercised earlier as a JS spike against the real TV, not yet
through this form on a device.
